refactor: extract IMedewerkersOverzichtService with page limit
- Extract business logic from controller into IMedewerkersOverzichtService
- Add MaxPages=10 cap to recursive pagination (prevents unbounded recursion)
- Register service in DI (TECH.md §6 compliance)
- Controller now only handles HTTP concerns
